+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0
+ - Also configure global `.npmrc` files.
+ - Provide additional guidance via the CLI output.
+ - Fix an error that occurs for Yarn 2 when the global config file is empty.
+
 ## 1.0.0
 
  - Initial Release